Change trace API to take &self instead of &mut self
The GC logic assumes calling `trace()` multiple times would visit the
same set of nodes.

Change the API to take `&self` to forbid write operations on `self`.

The main use of `trace` are all reading and writing to `CcBoxData`,
which uses interior mutability now. So it seems natural to drop `mut`
here.

This also makes the Trace interface easier to implement in certain
cases. For example:

    use external_crate::B;
    struct A(B(Cc<Box<dyn Trace>>));

    impl Trace for A {
        // B from external crate only exposes `as_ref()` without
        // `as_mut()` method.
    }