Add Trace::is_atomic to opt out the collector
Atomic types (i.e. do not have Cc dependencies) cannot opt-out the collector as
they cannot cause cycles.

This is similar to the "Green" color in the paper. But more aggressive: the
collector does not try to color or free them even if they are referred by
a container type. This is because the `Drop` implementation of the container
type would do the "free" thing. The paper was targeting JVM, which (I guess)
does not have a separate `Drop` implementation, and `Drop` is implemented
directly by the collector. So the collector there cannot skip visiting atomic
objects.

Unlike the paper, the "is_atomic" property (or the "Green" color) is not a
static property set upon instance creation, but a dynamic function that might
return different values at different times. This is potentially useful as a
future optimization. For example, CPython can dynamically decides whether
instances of container types opt in the collector or not by checking instance
types stored in the container.